requirements.txt
# 루시아 봇 의존성 - 고급 자연어 AI 어시스턴트
# 완전 업그레이드된 디스코드 봇을 위한 전체 의존성

# 핵심 라이브러리
discord.py>=2.4.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# 데이터베이스 및 영구 저장
aiosqlite>=0.20.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# 웹 스크래핑 및 HTTP 요청
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0

# 자연어 처리 고급 기능
nltk>=3.8.0
textblob>=0.17.0
spacy>=3.7.0
transformers>=4.35.0
sentence-transformers>=2.2.0

# 벡터 검색 및 임베딩
faiss-cpu>=1.7.4
chromadb>=0.4.0
langchain>=0.1.0
langchain-community>=0.0.10
openai>=1.0.0

# 뉴스 및 데이터 수집
feedparser>=6.0.10
newspaper3k>=0.2.8
scrapy>=2.11.0
tweepy>=4.14.0

# 고급 텍스트 처리
tiktoken>=0.5.0
unstructured>=0.10.0
pypdf>=3.17.0
python-docx>=1.1.0

# 이미지 처리 및 생성 (초고급)
Pillow>=10.1.0
opencv-python>=4.8.0
matplotlib>=3.8.0
numpy>=1.24.0
face-recognition>=1.3.0
pytesseract>=0.3.10
qrcode>=7.4.2

# 음성 처리 (초고급)
pydub>=0.25.0
speech-recognition>=3.10.0
gtts>=2.4.0
pyttsx3>=2.90
pyaudio>=0.2.11

# 웹 검색 및 API
google-search-results>=2.4.0
wikipedia>=1.4.0
youtube-dl>=2021.12.17

# 비동기 처리 및 성능 최적화
aiofiles>=24.1.0
asyncio-throttle>=1.0.2
redis>=5.0.0
celery>=5.3.0

# 로깅 및 모니터링
colorlog>=6.8.0
psutil>=5.9.0
prometheus-client>=0.19.0

# 보안 강화
cryptography>=41.0.0
bcrypt>=4.1.0
pyjwt>=2.8.0

# 데이터 분석 및 시각화 (초고급)
pandas>=2.1.0
plotly>=5.17.0
seaborn>=0.13.0
wordcloud>=1.9.2
scikit-learn>=1.3.0

# 게임 및 엔터테인먼트
chess>=1.10.0
python-chess>=1.999
pygame>=2.5.0

# 번역 및 다국어 지원
googletrans>=4.0.0
deep-translator>=1.11.4
langdetect>=1.0.9

# 코드 분석 및 보안
bandit>=1.7.5
pylint>=3.0.0
mypy>=1.7.0
radon>=6.0.1

# 스케줄링 및 작업 관리
apscheduler>=3.10.4
croniter>=2.0.1

# 고급 웹 스크래핑
playwright>=1.40.0
httpx>=0.25.0

# 유틸리티
python-dateutil>=2.8.0
pytz>=2023.3
schedule>=1.2.0
humanize>=4.8.0
emoji>=2.8.0
python-magic>=0.4.27

# 개발 및 테스트 (선택적)
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0